+++ src/CharactersServer/Characters/IronmanServer.h
@@ -20,7 +20,7 @@ public:
 
 
 private:
-    void moveRight(int distance, int vel, Box *boxOfEnemy, bool isGrounded);
+    void moveRight(int distance, float vel, Box *boxOfEnemy, bool isGrounded);
     void moveLeft(int distance, float vel, Box *boxOfEnemy, bool isGrounded);
     void stand();
     void update(int distance, int posContrincante, actions_t actionRecieved, Box *boxEnemy) override;
